refactor: move login page styles
