+++ src/pages/login/Input.jsx
@@ -1,47 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import {
-  Input,
-  InputGroup,
-  InputGroupText,
-  InputGroupAddon,
-  FormGroup,
-  FormFeedback
-} from "reactstrap";
-
-import styled from "styled-components";
-
-const inputStyle = {
-  border: 0,
-  color: "#b2b2b2",
-  height: 56,
-  lineHeight: 1.875,
-  borderRadius: 0,
-  borderTopRightRadius: 5,
-  borderBottomRightRadius: 5
-};
-
-const inputGroupStyle = {
-  border: 0,
-  background: "#fff",
-  borderTopLeftRadius: 5,
-  borderBottomLeftRadius: 5,
-  paddingLeft: 15
-};
-
-const inputGroupTextStyle = {
-  background: 0,
-  border: 0,
-  color: "#b2b2b2"
-};
-
-const Icon = styled.img`
-  height: 18px;
-`;
+import { Input, InputGroup, FormGroup, FormFeedback } from "reactstrap";
+import { inputStyle } from "./Input.styles";
+import PreIcon from "./input/PreIcon";
 
 const StyledInput = ({
   type,
+  name,
   icon,
   placeholder,
   value,
@@ -50,28 +16,27 @@ const StyledInput = ({
 }) => (
   <FormGroup>
     <InputGroup>
-      <InputGroupAddon style={inputGroupStyle} addonType="prepend">
-        <InputGroupText style={inputGroupTextStyle}>
-          <Icon src={icon} />
-        </InputGroupText>
-      </InputGroupAddon>
+      <PreIcon icon={icon} />
       <Input
         bsSize="lg"
         style={inputStyle}
         type={type}
-        name={type}
+        name={name}
         placeholder={placeholder}
         value={value}
         onChange={onChange}
         invalid={!!isInvalid}
       />
+      {isInvalid ? <FormFeedback invalid>{isInvalid}</FormFeedback> : null}
     </InputGroup>
   </FormGroup>
 );
 
 StyledInput.propTypes = {
-  onChange: PropTypes.func,
-  value: PropTypes.string,
+  onChange: PropTypes.func.isRequired,
+  value: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired,
+  isInvalid: PropTypes.string,
   type: PropTypes.string
 };
 StyledInput.defaultProps = {
