+++ src/pages/login/Input.styles.jsx
@@ -1,0 +1,11 @@
+const inputStyle = {
+  border: 0,
+  color: "#b2b2b2",
+  height: 56,
+  lineHeight: 1.875,
+  borderRadius: 0,
+  borderTopRightRadius: 5,
+  borderBottomRightRadius: 5
+};
+
+export { inputStyle };
