+++ src/pages/login/input/PreIcon.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import { InputGroupText, InputGroupAddon } from "reactstrap";
+import { Icon, inputGroupStyle, inputGroupTextStyle } from "./PreIcon.styles";
+
+const PreIcon = ({ icon }) => (
+  <InputGroupAddon style={inputGroupStyle} addonType="prepend">
+    <InputGroupText style={inputGroupTextStyle}>
+      <Icon src={icon} />
+    </InputGroupText>
+  </InputGroupAddon>
+);
+
+export default PreIcon;
