+++ src/pages/login/input/PreIcon.styles.jsx
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+const inputGroupStyle = {
+  border: 0,
+  background: "#fff",
+  borderTopLeftRadius: 5,
+  borderBottomLeftRadius: 5,
+  paddingLeft: 15
+};
+
+const inputGroupTextStyle = {
+  background: 0,
+  border: 0,
+  color: "#b2b2b2"
+};
+
+const Icon = styled.img`
+  height: 18px;
+`;
+
+export { Icon, inputGroupStyle, inputGroupTextStyle };
